rearrange again to allow usage on server and client
