+++ sources/index.ts
@@ -1,21 +1,6 @@
-import { ExporterMap, Source } from '../types'
-import { BafinExporter, BafinStimmrechte } from './bafin-stimmrechte'
-import { ArzneiEngpassExporter, Lieferengpass } from './bfg-lieferengpass'
-import { Gemeindeverzeichnis, GVExporter } from './destatis-gemeindeverzeichnis'
+import { Source } from '../types'
+import { BafinStimmrechte } from './bafin-stimmrechte'
+import { Lieferengpass } from './bfg-lieferengpass'
+import { Gemeindeverzeichnis } from './destatis-gemeindeverzeichnis'
 
 export const sources: Array<Source> = [BafinStimmrechte, Lieferengpass, Gemeindeverzeichnis]
-
-export const exporters: Array<ExporterMap> = [
-  {
-    id: BafinStimmrechte.id,
-    exporter: BafinExporter
-  },
-  {
-    id: Lieferengpass.id,
-    exporter: ArzneiEngpassExporter
-  },
-  {
-    id: Gemeindeverzeichnis.id,
-    exporter: GVExporter
-  }
-]
